% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde.r
\docType{package}
\name{cde-package}
\alias{cde-package}
\title{cde: Download Data from the Catchment Data Explorer Website}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Facilitates searching, download and plotting of Water Framework Directive (WFD) reporting data for all waterbodies within the UK Environment Agency area. The types of data that can be downloaded are: WFD status classification data, Reasons for Not Achieving Good (RNAG) status, objectives set for waterbodies, measures put in place to improve water quality and details of associated protected areas. The site accessed is \url{https://environment.data.gov.uk/catchment-planning/}. The data are made available under the Open Government Licence v3.0 \url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/cde}
  \item \url{https://github.com/ropensci/cde}
  \item Report bugs at \url{https://github.com/ropensci/cde/issues}
}

}
\author{
\strong{Maintainer}: Rob Briers \email{r.briers@napier.ac.uk} (\href{https://orcid.org/0000-0003-0341-1203}{ORCID})

}
\keyword{internal}
