% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde.r
\docType{package}
\name{cde}
\alias{cde}
\alias{cde-package}
\title{cde: Extract and plot Water Framework Directive (WFD) waterbody status data
from the Environment Agency Catchment Data Explorer (CDE) website}
\description{
The cde package allows searching, download and plotting of the WFD
waterbody status classification data for all waterbodies within the
Environment Agency area.
}
\details{
The website that is accessed is:
(http://environment.data.gov.uk/catchment-planning/).
The data are made available under the Open Government Licence v3.0
(https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/)
}
