% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE.R
\name{CDE_sample}
\alias{CDE_sample}
\title{Function for conditional sampling - sample the CDE}
\usage{
CDE_sample(cde.obj, Ytest)
}
\arguments{
\item{cde.obj}{An hdrcde::cde object}

\item{Ytest}{Predicted value of testing data}
}
\description{
Function for conditional sampling - sample the CDE
}
