% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.r
\name{download_ea}
\alias{download_ea}
\title{Download EA Catchment Data}
\usage{
download_ea(col_value = NULL, column = NULL, type = NULL,
  startyr = NULL, endyr = NULL)
}
\arguments{
\item{col_value}{A string representing the features to be extracted. For
example to extract data for the whole of the Humber RBD, this would be 
"Humber". Must be an exact match to the values used in the EA database.
Use the \code{\link{search_sites}} function to check for matching values.}

\item{column}{The column to be searched. Possible options are 
\code{wbid} (waterbody id), \code{oc} (Operational Catchment), \code{mc} 
(Management Catchment) and \code{rbd} (River Basin District)}

\item{type}{Type of waterbody to be extracted. For Operational/Management 
catchment level or RBD level queries, the data can also be subset by 
waterbody type. Possible values are \code{River}, \code{Lake}, 
\code{Groundwater}, \code{Transitional} or \code{Coastal}.}

\item{startyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. If only \code{startyr} is 
specified this extracts for a particular year.}

\item{endyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. The \code{endyr} should
only be specified if \code{startyr} is also included, otherwise it
is ignored.}
}
\value{
A data frame containing the classifcation details for the 
specified combination of column and value.
}
\description{
Downloads classification data from EA Catchment Data
Explorer site. Data can be downloaded by specifying waterbody id 
(\code{wbid}), Management Catchment (\code{mc}), Operational 
Catchment (\code{oc}) or River Basin District (\code{rbd}).
Start year (\code{startyr}) and end year (\code{endyr}) allow 
specific timeranges to be downloaded.
For Management Catchment (\code{mc}), Operational 
Catchment (\code{oc}) or River Basin District (\code{rbd}) level
downloads, waterbody type can also be specified.
}
\examples{

}
