% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE.R
\name{fit_cde}
\alias{fit_cde}
\title{Function for conditional sampling - create cde object}
\usage{
fit_cde(
  Yhat,
  resid,
  method_param = 2,
  xden_param = "uniform",
  xmar = 50,
  ymar = 100
)
}
\arguments{
\item{Yhat}{True value of calibration data}

\item{resid}{Predicted value of calibration data}

\item{method_param}{The argument `method` for function `hdrcde::cde.bandwidths`}

\item{xden_param}{The argument `xden` for function `hdrcde::cde.bandwidths`}

\item{xmar}{Number of bins for conditional density estimation}

\item{ymar}{Number of bins for conditional density estimation}
}
\description{
Function for conditional sampling - create cde object
}
