% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pa.r
\name{get_pa}
\alias{get_pa}
\title{Retrieve Protected Area Information}
\usage{
get_pa(ea_name = NULL, column = NULL)
}
\arguments{
\item{ea_name}{A string representing the description (name) of the
features to be extracted. For example to extract data for the whole of
the Humber RBD, this would be "Humber"; also see examples. Must be an
exact match to the values used in the EA database.
Use the \code{\link{search_names}} function to search for specific values.}

\item{column}{The column to be searched. Possible options are
\code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC}
(Management Catchment) and \code{RBD} (River Basin District)}
}
\value{
A data frame containing the details of the Protected Areas 
associated with the waterbodies.
}
\description{
Retrieves details of Protected Areas associated with 
waterbodies, catchments or River Basin Districts from the EA 
Catchment Data Explorer site.
Data can be retrieved by specifying waterbody id
(\code{WBID}), Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}).
Start year (\code{startyr}) and end year (\code{endyr}) allow
specific timeranges to be downloaded.
For Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}) level
downloads, waterbody \code{Type} can also be specified to allow
extraction of specific waterbody types (River, Lake etc).
}
\examples{
# get protected areas associated with waterbody GB112071065700
\dontrun{get_pa("GB112071065700", "WBID")}

# get the protected areas associated with Lakes in the Humber RBD
\dontrun{get_pa("Humber", "RBD", type = "Lake")}

# get the protected areas associated with Rivers in the Avon Warwickshire
# Operational Catchment
\dontrun{get_pa("Avon Warwickshire", "MC", type = "River")}

}
