% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rnag.r
\name{get_rnag}
\alias{get_rnag}
\title{Retrieve Reasons for Not Achieving Good Status}
\usage{
get_rnag(col_value = NULL, column = NULL, startyr = NULL,
  endyr = NULL, type = NULL)
}
\arguments{
\item{col_value}{A string representing the description (name) of the
features to be extracted. For example to extract data for the whole of
the Humber RBD, this would be "Humber"; also see examples. Must be an
exact match to the values used in the EA database.
Use the \code{\link{search_names}} function to search for specific values.}

\item{column}{The column to be searched. Possible options are
\code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC}
(Management Catchment) and \code{RBD} (River Basin District)}

\item{startyr}{The data can be extracted for specific years using the
\code{startyr} and \code{endyr} arguments. If only \code{startyr} is
specified this extracts for a particular year. If no years are specified
all years are returned.}

\item{endyr}{The data can be extracted for specific years using the
\code{startyr} and \code{endyr} arguments. The \code{endyr} should
only be specified if \code{startyr} is also included, otherwise it
is ignored and all years are returned.}

\item{type}{Type of waterbody to be extracted. For Operational/Management
catchment level or RBD level queries, the data can also be subset by
waterbody type. Possible values are \code{River}, \code{Lake},
\code{GroundWaterBody}, \code{TransitionalWater} or \code{CoastalWater}.}
}
\value{
A data frame containing the details of the Reasons for Not 
Achieving Good Status for the specified combination of column, value, 
dates and type.
}
\description{
Retrieves details of Reasons for Not Achieving Good (RNAG)
status from EA Catchment Data Explorer site.
Data can be retrieved by specifying waterbody id
(\code{WBID}), Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}).
Start year (\code{startyr}) and end year (\code{endyr}) allow
specific timeranges to be downloaded.
For Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}) level
downloads, waterbody \code{Type} can also be specified to allow
extraction of specific waterbody types (River, Lake etc).
}
\examples{
# get RNAG issues identified for waterbody GB112071065700
\dontrun{get_rnag("GB112071065700", "WBID")}

# get the RNAG issues for Lakes in the Humber RBD, between
# 2012 and 2014
\dontrun{get_rnag("Humber", "RBD", startyr = 2012, endyr = 2014, type = "Lake")}

# get the RNAG issues for Rivers in the Avon Warwickshire
# Operational Catchment in 2011
\dontrun{get_rnag("Avon Warwickshire", "MC", startyr = 2011, type = "River")}
}
