% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_new.r
\name{plot_status}
\alias{plot_status}
\title{Plot Status Summary}
\usage{
plot_status(x, scheme)
}
\arguments{
\item{x}{An object of class `cde_df`to be plotted}

\item{scheme}{Which colour scheme to use with plots; defaults to a viridis
based scheme (\code{"vir"} but can also choose to use the colours specified
in the WFD document by specifying as \code{"wfd"}.}
}
\value{
A (stacked) barplot of the percentage of waterbodies within the
specified area of different status values represented as different colours
depending on the scheme specified.
}
\description{
Produces a (stacked) percentage barplot of waterbody
status information for a given set of data (MC, OC or RBD)
}
