% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde-class_and_methods.r
\name{print.cde_df}
\alias{print.cde_df}
\title{Print method for cde_df}
\usage{
\method{print}{cde_df}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cde_df}.}

\item{...}{Other arguments passed on to individual methods. None 
implemented at present.}
}
\description{
Custom \code{print} method for objects of class \code{cde_df}.
Formats output to fit current width of console, keeping full column names 
but truncating row values as required.
}
