% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_sites.r
\name{search_sites}
\alias{search_sites}
\title{Search EA site database}
\usage{
search_sites(string = NULL, column = NULL)
}
\arguments{
\item{string}{The search string to be matched. Will match whole or partial
strings in the column values.}

\item{column}{The column to be searched. Possible options are 
\code{name}, \code{OC} (Operational Catchment), \code{MC} 
(Management Catchment) and \code{RBD} (River Basin District)}
}
\value{
A data frame containing the details of all the sites that match
the search string (full or partial matches) in the column specified.
}
\description{
Searches the listing of EA monitoring sites to find rows
that contain the string provided. Can search by name (\code{name}), 
Management Catchment (\code{MC}), Operational Catchment (\code{OC}) 
or River Basin District (\code{RBD}).
}
\examples{
search_sites("Avon", "OC")
search_sites("Tweed", "RBD")
}
