% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_plot.r
\name{status_plot}
\alias{status_plot}
\title{Plot Status Summary}
\usage{
status_plot(col_value = NULL, column = NULL,
  element = "Overall Water Body", startyr = NULL, endyr = NULL,
  type = NULL)
}
\arguments{
\item{col_value}{A string representing the description (name) of the 
features to be extracted. For example to extract data for the whole of 
the Humber RBD, this would be "Humber"; also see examples. Must be an 
exact match to the values used in the EA database.
Use the \code{\link{search_sites}} function to search for specific values.}

\item{column}{The column to be searched. Possible options are 
\code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC} 
(Management Catchment) and \code{RBD} (River Basin District)}

\item{element}{The WFD quality element to be extracted. Defaults to 'Overall
status'. See Vignette for possible values.}

\item{startyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. If only \code{startyr} is 
specified this extracts for a particular year. If no years are specified 
all years are returned.}

\item{endyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. The \code{endyr} should
only be specified if \code{startyr} is also included, otherwise it
is ignored.}

\item{type}{Type of waterbody to be extracted. For Operational/Management 
catchment level or RBD level queries, the data can also be subset by 
waterbody type. Possible values are \code{River}, \code{Lake}, 
\code{Groundwater}, \code{Transitional} or \code{Coastal}.}
}
\value{
A stacked barplot of the percentage of waterbodies within the 
specified area of different status values, with standard colour scheme.
}
\description{
Produces a stacked proportional barplot of waterbody
status information for a given set of data (MC, OC or RBD)
}
