% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfd_class.r
\name{wfd_class}
\alias{wfd_class}
\title{Retrieve WFD Classification Information}
\usage{
wfd_class(col_value = NULL, column = NULL, element = "Overall status",
  startyr = NULL, endyr = NULL, type = NULL)
}
\arguments{
\item{col_value}{A string representing the description (name) of the 
features to be extracted. For example to extract data for the whole of 
the Humber RBD, this would be "Humber"; also see examples. Must be an 
exact match to the values used in the EA database.
Use the \code{\link{search_sites}} function to search for specific values.}

\item{column}{The column to be searched. Possible options are 
\code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC} 
(Management Catchment) and \code{RBD} (River Basin District)}

\item{element}{The WFD quality element to be extracted. Defaults to 'Overall
status'. See Vignette for possible values.}

\item{startyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. If only \code{startyr} is 
specified this extracts for a particular year. If no years are specified 
all years are returned.}

\item{endyr}{The data can be extracted for specific years using the 
\code{startyr} and \code{endyr} arguments. The \code{endyr} should
only be specified if \code{startyr} is also included, otherwise it
is ignored.}

\item{type}{Type of waterbody to be extracted. For Operational/Management 
catchment level or RBD level queries, the data can also be subset by 
waterbody type. Possible values are \code{River}, \code{Lake}, 
\code{Groundwater}, \code{Transitional} or \code{Coastal}.}
}
\value{
A data frame containing the classifcation details for the 
specified combination of column, value, element and dates.
}
\description{
Retrieves classification data from EA Catchment Data
Explorer site. Data can be retrieved by specifying waterbody id 
(\code{WBID}), Management Catchment (\code{MC}), Operational 
Catchment (\code{OC}) or River Basin District (\code{RBD}).
Start year (\code{startyr}) and end year (\code{endyr}) allow 
specific timeranges to be downloaded.
For Management Catchment (\code{MC}), Operational 
Catchment (\code{OC}) or River Basin District (\code{RBD}) level
downloads, waterbody \code{Type} can also be specified to allow
extraction of specific waterbody types (River, Lake etc).
}
