% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.r
\name{zip_download}
\alias{zip_download}
\title{Download Zipfile and extract csv}
\usage{
zip_download(download_url)
}
\arguments{
\item{download_url}{A string representing the url to download the
zip file from.}
}
\description{
Downloads zipfile from specified url, unzips to
csv file and reads csv into dataframe.
}
