% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde.r
\name{cde}
\alias{cde}
\title{cde: Download Water Framework Directive (WFD) data from the 
Environment Agency Catchment Data Explorer (CDE) website.}
\description{
Facilitates searching and download of the WFD-related
data for all waterbodies within the Environment Agency area (i.e. England).
The types of data that can be downloaded are: WFD status classification 
data, Reasons for Not Achieving Good (RNAG) status, objectives set for 
waterbodies, and details of associated protected areas. Default plots can 
also be produced from thedata downloaded (form of plot depends on data type).
}
\details{
The website that is accessed is: 
\url{https://environment.data.gov.uk/catchment-planning/}.
The data accessed by and included within the package are made available 
under the Open Government Licence v3.0 
\url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}
}
