% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.r
\name{get_status}
\alias{get_status}
\title{Retrieve WFD Status Classification Data}
\usage{
get_status(ea_name = NULL, column = NULL,
  level = "Overall Water Body", startyr = NULL, endyr = NULL,
  type = NULL)
}
\arguments{
\item{ea_name}{A string representing the description (\code{name} for 
\code{OC}, \code{MC} or \code{RBD} level downloads or \code{WBID} for 
individual waterbodies) of the features to be extracted. For example 
to extract data for the whole of the Humber RBD, this would be "Humber"; 
also see examples. Must be an exact match to the values used in the EA 
database. Use the \code{\link{search_names}} function to search for 
specific values.}

\item{column}{The column to be searched. Possible options are
\code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC}
(Management Catchment) and \code{RBD} (River Basin District)}

\item{level}{The level within the WFD quality classification elements that 
objectives have been set at. For full details of the hierarchy of elements 
within the classification used, see \url{https://environment.data.gov.uk/catchment-planning/help#help-classification-hierarchy}.

Defaults to 'Overall Water Body'. Possible values for the different levels 
retrieved by the function are shown below.
\tabular{ccc}{
\strong{Level 1} \tab \strong{Level 2} \tab \strong{Level 4}\cr
Ecological \tab Biological quality elements \tab Overall Water Body\cr
Chemical \tab Chemical Status element \tab -\cr
Quantitative \tab Hydromorphological Supporting Elements \tab -\cr
- \tab Other Substances \tab -\cr
- \tab Physico-chemical quality elements \tab -\cr
- \tab Priority hazardous substances \tab -\cr
- \tab Priority substances \tab -\cr
- \tab Quantitative Status element \tab - \cr
- \tab Specific pollutants \tab -\cr
- \tab Supporting elements \tab -\cr
}}

\item{startyr}{The data can be extracted for specific years using the
\code{startyr} and \code{endyr} arguments. If only \code{startyr} is
specified this extracts for a particular year. If no years are specified
all years are returned.}

\item{endyr}{The data can be extracted for specific years using the
\code{startyr} and \code{endyr} arguments. The \code{endyr} should
only be specified if \code{startyr} is also included, otherwise an 
error is returned.}

\item{type}{Type of waterbody to be extracted. For Operational/Management
catchment level or RBD level queries, the data can also be subset by
waterbody type. Possible values are \code{River}, \code{Lake},
\code{GroundWaterBody}, \code{TransitionalWater} or \code{CoastalWater}.}
}
\value{
An object of class \code{cde_df} containing the classification 
details for the specified combination of criteria.
For details of the meaning of the the different columns returned, 
see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
}
\description{
Retrieves WFD Status classification data from EA Catchment 
Data Explorer site. Data can be retrieved by specifying waterbody id
(\code{WBID}), Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}).
Start year (\code{startyr}) and end year (\code{endyr}) allow
specific timeranges to be downloaded.
For Management Catchment (\code{MC}), Operational
Catchment (\code{OC}) or River Basin District (\code{RBD}) level
downloads, waterbody \code{type} can also be specified to allow
extraction of specific waterbody types (River, Lake etc).
}
\examples{
# get Overall Water Body status classification for waterbody GB520804714300
get_status(ea_name="GB520804714300", column="WBID")

# get status class based on Priority substances for waterbody GB520804714300
get_status(ea_name="GB520804714300", column="WBID", level="Priority substances")

# get the Overall Water Body status of Lakes in the Humber RBD, between
# 2012 and 2014
get_status(ea_name="Humber", column="RBD", startyr=2012, endyr=2014, type="Lake")

# get the Overall Water Body status for Rivers in the Avon Warwickshire
# Operational Catchment in 2011
get_status(ea_name="Avon Warwickshire", column="MC", startyr=2011, type="River")

}
