% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde-class.r
\name{plot.cde_df}
\alias{plot.cde_df}
\title{Plot method for cde_df output}
\usage{
\method{print}{cde_df}(x, scheme = "vir", ...)
}
\arguments{
\item{x}{An object of class `cde_df` to be plotted}

\item{scheme}{Which colour scheme to use with plots; defaults to a viridis
based scheme (\code{"vir"} but can also choose to use the colours specified
in the WFD document by specifying as \code{"wfd"}.}
}
\description{
Default plots of the output of calls to `get...` functions.
For `status` and `objectives` produces a (stacked) percentage barplot of 
waterbody observed or predicted (objective) status information for a 
given set of data. 

For `rnag`, `measures` or `pa` produces a frequency
histogram. NEED TO INCLUDE LINK TO EA classification site here. 
Plotting is only possible for MC, OC or RBD downloads.
}
